requirements.txt
django
pillow
django-admin-interface
django-crispy-forms
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
idna==2.10
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.5
djangorestframework